WIP: Update to device name as seen by Linux PC/SC. (SEE NOTES TO SELF)
This is pretty awful to get running on new JVMs on the Pi Zero; to be able
to use the JIT on ARMv6 chips, you need to use Java 8 or earlier.  If you
want to use a newer version of Java, you need to use the "Zero" JVM, which
is interpreted and _extremely slow_.

The slowness is so bad that it actually prevents the cards from being read
correctly.

I think the way forward with this is going to be getting Kotlin building
native code for the platform. That's a task in itself because of getting
Kotlin/Native up-to-date, and also getting it to accept extra headers (from
libpcsclite).